+++ MiniProgram/src/roomkit/pages/home.vue
@@ -169,6 +169,7 @@ handleInit();
 </script>
 
 <style>
+@import '@TUIRoom/assets/style/global.scss';
 @import '@TUIRoom/assets/style/black-theme.scss';
 @import '@TUIRoom/assets/style/white-theme.scss';
 /* * {