+++ MiniProgram/wxmini_dev.sh
@@ -1,4 +1,5 @@
 cd dist/dev/mp-weixin
+cd unpackage/dist/dev/mp-weixin
 npm init -y
 npm i @tencentcloud/trtc-component-wx
 cd roomkit