+++ MiniProgram/wxmini_prod.sh
@@ -1,4 +1,5 @@
 cd dist/build/mp-weixin
+cd unpackage/dist/build/mp-weixin
 npm init -y
 npm i @tencentcloud/trtc-component-wx
 cd roomkit