+++ MiniProgram/CHANGELOG.md
@@ -1,8 +1,37 @@
+## 2023.12.18@1.7.0
+
+**Feature**
+
+- 升级 [@tencentcloud/tuiroom-engine-wx](https://www.npmjs.com/package/@tencentcloud/tuiroom-engine-wx) 到 v1.7.0 版本，详情请查看 [发布日志](https://cloud.tencent.com/document/product/1690/89361)；
+- 新增会中移交主持人功能。
+- 新增结束会议及踢出房间确认弹窗。
+- 小程序新增视频流懒加载功能。
+- 主持人邀请成员开启视频、麦克风增加消息提示。
+- 移交主持人新增错误信息提示。
+- 优化退房性能。
+
+**Bug Fixed**
+
+- 修复默认头像在小程序无法加载的问题。
+- 修复观众端禁止文字聊天区域的展示交互。
+- 修复上台发言模式下，第二次上麦打不开麦克风问题。
+- 修复被踢出房间 live-pusher 被提前销毁问题。
+- 修复小程序成员列表功能点击失效问题。
+- 修复设备测试区域，切换扬声器和播放设备不生效的问题。
+- 修复移动端小程序镜像设置设置不生效问题。
+- 修复移动端小程序成员管理，操作最后一个成员时，样式异常问题。
+- 修复移动端小程序下 dialog 文案及字体颜色错误问题。
+- 修复移动端小程序点击新的成员管理弹窗，旧弹窗不会被关闭问题。
+- 修复小程序收不到 IM 加入群组事件，无法展示 userName 及无法正常移交主持人问题。
+- 修复会中切换前后置摄像头，然后关闭摄像头后再打开，前后置摄像头状态显示错误问题。
+- 修复小程序首次渲染偶现黑屏问题。
+- 增加小程序回退图标按钮区域大小，修复手指点击无响应问题。
+
 ## 2023.11.14@1.6.1
 
 **Feature**
 
-- 升级 [@tencentcloud/tuiroom-engine-js](https://www.npmjs.com/package/@tencentcloud/tuiroom-engine-js) 到 v1.6.1 版本，详情请查看 [发布日志](https://cloud.tencent.com/document/product/1690/89361)；
+- 升级 [@tencentcloud/tuiroom-engine-wx](https://www.npmjs.com/package/@tencentcloud/tuiroom-engine-wx) 到 v1.6.1 版本，详情请查看 [发布日志](https://cloud.tencent.com/document/product/1690/89361)；
 - 新增支持切换黑白主题，更改 svg 的引入方式；
 
 **Bug Fixed**